+++ example/lib/splash/splash_screen_view_model.dart
@@ -1,3 +1,4 @@
+
 import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/stateful/base_view_model.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -38,4 +39,16 @@ class SplashScreenViewModel
 
   @override
   void viewDisposed() {}
+
+  @override
+  void viewResumed() {}
+
+  @override
+  void viewInactive() {}
+
+  @override
+  void viewPaused() {}
+
+  @override
+  void viewSuspending() {}
 }