+++ lib/presentation/ui/mvvm/stateful/base_view_model.dart
@@ -1,34 +1,59 @@
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/utils/base_view_model_utils.dart';
+
 import 'base_model.dart';
 import 'base_view.dart';
 
-abstract class BaseViewModel<M extends BaseModel, V extends BaseView<M>> {
-  V _view;
-  M _model;
+abstract class BaseViewModel<M extends BaseModel, V extends BaseView<M>>
+    extends BaseViewModelUtils {
+  BaseViewModel(V view) : super(view, view.model);
 
-  V get view => _view;
-  M get model => _model;
+  /// Getting the view in MVVM
+  V get view => viewUtils as V;
 
-  BaseViewModel(this._view) {
-    _model = _view.model;
-    init();
-  }
+  /// Getting the model in MVVM
+  M get model => modelUtils as M;
 
+  /// Callback event during screen start
   void init() {
     model.viewCallbacks.viewCreatedCallback(viewCreated);
     model.viewCallbacks.viewRefreshCallback(viewRefresh);
     model.viewCallbacks.viewInitStateCallback(initState);
     model.viewCallbacks.viewDisposedCallback(viewDisposed);
+    model.viewCallbacks.viewResumedCallback(viewResumed);
+    model.viewCallbacks.viewInactiveCallback(viewInactive);
+    model.viewCallbacks.viewPausedCallback(viewPaused);
+    model.viewCallbacks.viewSuspendingCallback(viewSuspending);
     model.loadingShow.addCallback(loadingShow);
     model.loadingHide.addCallback(loadingHide);
   }
 
+  /// Event callback when the view already created.
   void viewCreated() {}
 
+  /// Event callback when need to refresh the view.
   void viewRefresh() {}
 
+  /// Event callback when initState finished.
   void initState() {}
 
+  /// Event callback when the view disposed.
   void viewDisposed() {}
+
+  /// Event callback when the application is visible and responding to user input.
+  void viewResumed() {}
+
+  /// Event callback when the application is in an inactive state and is not receiving user input.
+  void viewInactive() {}
+
+  /// Event callback when the application is not currently visible to the user, not responding to
+  /// user input, and running in the background.
+  void viewPaused() {}
+
+  /// Event callback when the application is in this state, the engine will not call the
+  /// [Window.onBeginFrame] and [Window.onDrawFrame] callbacks.
+  /// Android apps in this state should assume that they may enter the
+  /// [suspending] state at any time.
+  void viewSuspending() {}
 
   void loadingShow() {
     model.loading = true;