+++ lib/presentation/ui/mvvm/stateless/base_model.dart
@@ -1,4 +1,5 @@
-import '../utils/base_model_utils.dart';
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/utils/base_model_utils.dart';
+
 import 'base_view.dart';
 import 'base_view_model.dart';