+++ lib/presentation/ui/mvvm/stateless/base_view_model.dart
@@ -1,21 +1,23 @@
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/utils/base_view_model_utils.dart';
+
 import 'base_model.dart';
 import 'base_view.dart';
 
-abstract class BaseViewModel<M extends BaseModel, V extends BaseView<M>> {
-  V _view;
-  M _model;
+abstract class BaseViewModel<M extends BaseModel, V extends BaseView<M>>
+    extends BaseViewModelUtils {
+  BaseViewModel(V view) : super(view, view.model);
 
-  V get view => _view;
-  M get model => _model;
+  /// Getting the view in MVVM
+  V get view => viewUtils as V;
 
-  BaseViewModel(this._view) {
-    _model = _view.model;
-    init();
-  }
+  /// Getting the model in MVVM
+  M get model => modelUtils as M;
 
+  /// Callback event during screen start
   void init() {
     model.viewCallbacks.viewCreatedCallback(viewCreated);
   }
 
+  /// Event callback when the view already created.
   void viewCreated() {}
 }