+++ lib/presentation/ui/mvvm/utils/base_view_utils.dart
@@ -3,8 +3,9 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/utils/base_model_utils.dart';
 
-class BaseViewUtils {
+class BaseViewUtils<M extends BaseModelUtils> {
   Future<T> navigateTo<T extends Object>(
       BuildContext context, Widget widget, bool clear,
       {bool material = true}) {